third_party/WebKit/LayoutTests/fast/forms/month-multiple-fields/month-multiple-fields-ax-aria-attributes-expected.txt

This test checks aria-help attribute of fields in multiple fields week input UI.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Non-empty value
PASS focusedElementDescription() is "Month, AXValueDescription: October, intValue:10, range:1-12"
PASS focusedElementDescription() is "Year, AXValueDescription: 2012, intValue:2012, range:1-275760"
Empty value
PASS focusedElementDescription() is "Year, AXValueDescription: ----, intValue:0, range:1-275760"
PASS focusedElementDescription() is "Month, AXValueDescription: ---------, intValue:0, range:1-12"

PASS successfullyParsed is true

TEST COMPLETE
